+++ lib/multistreamwrapper.js
@@ -1,0 +1,42 @@
+var streamPair = require('stream-pair');
+var Interactive = require('multistream-select').Interactive;
+var Select = require('multistream-select').Select
+
+var pair = streamPair.create();
+var msi = new Interactive();
+var mss = new Select();
+
+module.exports = function () {
+  mss.handle(pair.other);
+
+  mss.addHandler('/xxx/', function(ds) {
+    ds.on('data', function(chunk) {
+      console.log(chunk.toString());
+    });
+    ds.on('end', function() {
+
+    });
+  });
+
+  mss.addHandler('/wtf/', function(ds) {
+    ds.on('data', function(chunk) {
+      console.log(chunk.toString());
+      console.log('w');
+    });
+    ds.on('end', function() {
+      ds.end();
+    });
+  });
+
+
+  msi.handle(pair, function() {
+    msi.ls(function(err, result) {
+      if (err) {
+        return console.log(err)
+      }
+      console.log('results: ', result);
+    });
+  });
+
+  return msi
+}
