+++ npm-debug.log
@@ -1,18 +1,18 @@
 0 info it worked if it ends with ok
-1 verbose cli [ '/usr/local/bin/node', '/usr/local/bin/npm', 'run', 'test' ]
+1 verbose cli [ '/usr/local/bin/node', '/usr/local/bin/npm', 'run', 'dev' ]
 2 info using npm@3.4.0
 3 info using node@v4.1.1
-4 verbose run-script [ 'pretest', 'test', 'posttest' ]
-5 info lifecycle multicodec-doc@1.0.0~pretest: multicodec-doc@1.0.0
-6 silly lifecycle multicodec-doc@1.0.0~pretest: no script for pretest, continuing
-7 info lifecycle multicodec-doc@1.0.0~test: multicodec-doc@1.0.0
-8 verbose lifecycle multicodec-doc@1.0.0~test: unsafe-perm in lifecycle true
-9 verbose lifecycle multicodec-doc@1.0.0~test: PATH: /Users/josep/.npm-packages/lib/node_modules/npm/bin/node-gyp-bin:/Users/josep/Projects/github/ipfs/multicodec-doc/node_modules/.bin:/usr/local/bin:/usr/local/sbin:/Users/josep/bin:/usr/bin:/bin:/usr/sbin:/sbin:/usr/local/MacGPG2/bin
-10 verbose lifecycle multicodec-doc@1.0.0~test: CWD: /Users/josep/Projects/github/ipfs/multicodec-doc
-11 silly lifecycle multicodec-doc@1.0.0~test: Args: [ '-c', 'standard' ]
-12 silly lifecycle multicodec-doc@1.0.0~test: Returned: code: 1  signal: null
-13 info lifecycle multicodec-doc@1.0.0~test: Failed to exec test script
-14 verbose stack Error: multicodec-doc@1.0.0 test: `standard`
+4 verbose run-script [ 'predev', 'dev', 'postdev' ]
+5 info lifecycle multicodec-doc@1.0.0~predev: multicodec-doc@1.0.0
+6 silly lifecycle multicodec-doc@1.0.0~predev: no script for predev, continuing
+7 info lifecycle multicodec-doc@1.0.0~dev: multicodec-doc@1.0.0
+8 verbose lifecycle multicodec-doc@1.0.0~dev: unsafe-perm in lifecycle true
+9 verbose lifecycle multicodec-doc@1.0.0~dev: PATH: /Users/josep/.npm-packages/lib/node_modules/npm/bin/node-gyp-bin:/Users/josep/Projects/github/ipfs/multicodec-doc/node_modules/.bin:/usr/local/bin:/usr/local/sbin:/Users/josep/bin:/usr/bin:/bin:/usr/sbin:/sbin:/usr/local/MacGPG2/bin
+10 verbose lifecycle multicodec-doc@1.0.0~dev: CWD: /Users/josep/Projects/github/ipfs/multicodec-doc
+11 silly lifecycle multicodec-doc@1.0.0~dev: Args: [ '-c', 'beefy index:build.js --live ' ]
+12 silly lifecycle multicodec-doc@1.0.0~dev: Returned: code: 1  signal: null
+13 info lifecycle multicodec-doc@1.0.0~dev: Failed to exec dev script
+14 verbose stack Error: multicodec-doc@1.0.0 dev: `beefy index:build.js --live `
 14 verbose stack Exit status 1
 14 verbose stack     at EventEmitter.<anonymous> (/Users/josep/.npm-packages/lib/node_modules/npm/lib/utils/lifecycle.js:232:16)
 14 verbose stack     at emitTwo (events.js:87:13)
@@ -25,18 +25,18 @@
 15 verbose pkgid multicodec-doc@1.0.0
 16 verbose cwd /Users/josep/Projects/github/ipfs/multicodec-doc
 17 error Darwin 15.0.0
-18 error argv "/usr/local/bin/node" "/usr/local/bin/npm" "run" "test"
+18 error argv "/usr/local/bin/node" "/usr/local/bin/npm" "run" "dev"
 19 error node v4.1.1
 20 error npm  v3.4.0
 21 error code ELIFECYCLE
-22 error multicodec-doc@1.0.0 test: `standard`
+22 error multicodec-doc@1.0.0 dev: `beefy index:build.js --live `
 22 error Exit status 1
-23 error Failed at the multicodec-doc@1.0.0 test script 'standard'.
+23 error Failed at the multicodec-doc@1.0.0 dev script 'beefy index:build.js --live '.
 23 error Make sure you have the latest version of node.js and npm installed.
 23 error If you do, this is most likely a problem with the multicodec-doc package,
 23 error not with npm itself.
 23 error Tell the author that this fails on your system:
-23 error     standard
+23 error     beefy index:build.js --live
 23 error You can get their info via:
 23 error     npm owner ls multicodec-doc
 23 error There is likely additional logging output above.
